feat: added support for input_table and output tables
some data are stored  in arrays columns in the tickers json. I wanted a way to process them without having to create additional methods. Now when instantiating, it will fetch imformation based on the inpout table to create the dataframe and will fetch information on the output table to transform and upsert